AI assisted refactor
