+++ pixi.toml
@@ -15,3 +15,5 @@ pyyaml = ">=6.0.2,<7"
 openpyxl = ">=3.1.5,<4"
 pytest = ">=9.0.1,<10"
 python-build = ">=1.3.0,<2"
+pydantic = ">=2.12.4,<3"
+pandera = ">=0.26.1,<0.27"
