+++ pyproject.toml
@@ -23,6 +23,8 @@ dependencies = [
     "plotly>=5.15.0",
     "PyYAML>=6.0",
     "openpyxl>=3.0.0",
+    "pydantic>=2.0.0",
+    "pandera>=0.18.0",
 ]
 
 [project.scripts]
